LayoutTests/fast/forms/validationMessage-expected.txt

Test for validationMessage DOM property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS patternInput.validationMessage is 'pattern mismatch'
PASS requiredInput.validationMessage is 'value missing'
PASS requiredTextArea.validationMessage is 'value missing'
PASS emailInput.validationMessage is 'type mismatch'
PASS underInput.validationMessage is 'range underflow'
PASS overInput.validationMessage is 'range overflow'
PASS but.validationMessage is ''
PASS anoninput.validationMessage is ''
PASS happyFieldset.validationMessage is ''
PASS happySelect.validationMessage is ''
PASS successfullyParsed is true

TEST COMPLETE
